+++ geokey/users/migrations/0006_merge.py
@@ -1,0 +1,15 @@
+# -*- coding: utf-8 -*-
+from __future__ import unicode_literals
+
+from django.db import models, migrations
+
+
+class Migration(migrations.Migration):
+
+    dependencies = [
+        ('users', '0002_auto_20150904_1113'),
+        ('users', '0005_auto_20150825_0933'),
+    ]
+
+    operations = [
+    ]