+++ geokey/users/views.py
@@ -15,6 +15,7 @@ from rest_framework import status
 
 from geokey.applications.models import Application
 
+from geokey.core.views import ProjectContext
 from geokey.core.decorators import (
     handle_exceptions_for_ajax, handle_exceptions_for_admin
 )
@@ -26,7 +27,8 @@ from .serializers import (UserSerializer, UserGroupSerializer)
 from .forms import (
     UsergroupCreateForm,
     CustomPasswordChangeForm,
-    UserRegistrationForm
+    UserRegistrationForm,
+    UserForm
 )
 
 
@@ -60,10 +62,12 @@ class Index(TemplateView):
         django.http.HttpResponseRedirect
             to dashboard, if the user is signed in
         """
-        if request.user.is_anonymous():
-            return self.render_to_response(self.get_context_data)
-        else:
+        if not request.user.is_anonymous():
+            #     return self.render_to_response(self.get_context_data)
+            # else:
             return redirect('admin:dashboard')
+
+        return super(Index, self).get(request, *args, **kwargs)
 
 
 class Dashboard(LoginRequiredMixin, TemplateView):
@@ -101,31 +105,14 @@ class Dashboard(LoginRequiredMixin, TemplateView):
         }
 
 
-class UserGroupList(LoginRequiredMixin, TemplateView):
+class UserGroupList(LoginRequiredMixin, ProjectContext, TemplateView):
     """
     Displays a list of all user groups for a project.
     """
     template_name = 'users/usergroup_list.html'
 
-    @handle_exceptions_for_admin
-    def get_context_data(self, project_id):
-        """
-        Overwrites the method to add project to the context
 
-        Parameters
-        ----------
-        project_id : int
-            identifies the project in the data base
-
-        Returns
-        -------
-        dict
-        """
-        project = Project.objects.as_admin(self.request.user, project_id)
-        return super(UserGroupList, self).get_context_data(project=project)
-
-
-class UserGroupCreate(LoginRequiredMixin, CreateView):
+class UserGroupCreate(LoginRequiredMixin, ProjectContext, CreateView):
     """
     Displays the create user group page
     `/admin/projects/:project_id/usergroups/new/`
@@ -144,13 +131,7 @@ class UserGroupCreate(LoginRequiredMixin, CreateView):
         """
         project_id = self.kwargs['project_id']
 
-        context = super(
-            UserGroupCreate, self).get_context_data(**kwargs)
-
-        context['project'] = Project.objects.as_admin(
-            self.request.user, project_id
-        )
-        return context
+        return super(UserGroupCreate, self).get_context_data(project_id)
 
     def get_success_url(self):
         """
@@ -212,30 +193,12 @@ class UserGroupCreate(LoginRequiredMixin, CreateView):
         return self.render_to_response(context)
 
 
-class AdministratorsOverview(LoginRequiredMixin, TemplateView):
+class AdministratorsOverview(LoginRequiredMixin, ProjectContext, TemplateView):
     """
     Displays the list of administrators of the project
     `/admin/projects/:project_id/usergroups/:group_id/`
     """
     template_name = 'users/usergroup_admins.html'
-
-    @handle_exceptions_for_admin
-    def get_context_data(self, project_id):
-        """
-        Creates the request context for rendering the page, adds the project
-        to the context
-
-        Parameters
-        ----------
-        project_id : int
-            identifies the project in the data base
-
-        Return
-        ------
-        dict
-        """
-        project = Project.objects.as_admin(self.request.user, project_id)
-        return {'project': project}
 
 
 class UserGroupMixin(object):
@@ -384,41 +347,52 @@ class UserGroupDelete(LoginRequiredMixin, UserGroupMixin, TemplateView):
 
 class UserProfile(LoginRequiredMixin, TemplateView):
     """
-    Displays the user profile page
-    `/admin/profile`
+    Displays the user profile page.
     """
     template_name = 'users/profile.html'
 
     def post(self, request):
         """
-        Updates the user information
+        Updates user profile.
 
         Parameter
         ---------
         request : django.http.HttpRequest
-            Object representing the request.
+            Object representing the request
 
         Returns
         -------
         django.http.HttpResponse
             Rendered template
         """
-        user = request.user
+        user = User.objects.get(pk=request.user.pk)
+        form = UserForm(request.POST, instance=user)
 
-        user.display_name = request.POST.get('display_name')
-        new_email = request.POST.get('email')
+        if form.is_valid():
+            if form.has_changed():
+                user.display_name = form.cleaned_data['display_name']
+                user.email = form.cleaned_data['email']
+                user.save()
 
-        if user.email != new_email:
-            email = EmailAddress.objects.get(user=user, email=user.email)
-            email.change(request, new_email, confirm=True)
+                if user.email != request.user.email:
+                    try:
+                        EmailAddress.objects.get(user=user).change(
+                            request,
+                            user.email,
+                            confirm=True
+                        )
+                    except EmailAddress.DoesNotExist:
+                        EmailAddress.objects.create(
+                            user=user,
+                            email=user.email
+                        ).send_confirmation(request)
 
-            user.email = new_email
+                messages.success(request, 'Your profile has been updated.')
+                self.request.user = user
+            else:
+                messages.info(request, 'Your profile has not been edited.')
 
-        user.save()
-
-        context = self.get_context_data()
-        messages.success(request, 'The user information has been updated.')
-        return self.render_to_response(context)
+        return self.render_to_response(self.get_context_data(form=form))
 
 
 # ############################################################################
@@ -483,7 +457,7 @@ class UserGroup(APIView):
         project = Project.objects.as_admin(request.user, project_id)
         group = project.usergroups.get(pk=group_id)
         serializer = UserGroupSerializer(
-            group, data=request.DATA, partial=True)
+            group, data=request.data, partial=True)
 
         if serializer.is_valid():
             serializer.save()
@@ -518,7 +492,7 @@ class UserGroupUsers(APIView):
             Contains the serialised usergroup or an error message
         """
         project = Project.objects.as_admin(request.user, project_id)
-        user_id = request.DATA.get('userId')
+        user_id = request.data.get('userId')
 
         try:
             user = User.objects.get(pk=user_id)
@@ -659,7 +633,7 @@ class UserAPIView(CreateUserMixin, APIView):
                 status=status.HTTP_401_UNAUTHORIZED
             )
 
-        data = request.DATA
+        data = request.data
 
         serializer = UserSerializer(user, data=data, partial=True)
         if serializer.is_valid():
@@ -700,7 +674,7 @@ class UserAPIView(CreateUserMixin, APIView):
         rest_framework.response.Response
             Containing the user info or an error message
         """
-        data = request.DATA
+        data = request.data
         form = UserRegistrationForm(data)
         client_id = data.pop('client_id', None)
 
@@ -753,7 +727,7 @@ class ChangePasswordView(APIView):
             Empty response indicating success or error message
         """
         user = request.user
-        data = request.DATA
+        data = request.data
 
         if not user.is_anonymous():
             form = CustomPasswordChangeForm(user, data)