+++ CHANGELOG.md
@@ -4,6 +4,18 @@
 
 - Removes `/api/projects/:project_id/get-in-touch` endpoint. Also, notifications settings are now completely removed from GeoKey, therefore `contact_admins` method is no longer available on the Project model.
 
+## Version 0.7.2
+
+- Bug fixes:
+    - [#283](https://github.com/ExCiteS/geokey/issues/283) When deleting a review comment, the corresponding observation is only updated, when the status changes as well.
+    - [#289](https://github.com/ExCiteS/geokey/issues/289) Adds error messages to user registration formats
+    - [#290](https://github.com/ExCiteS/geokey/issues/290) Catches integrity errors when updating user profiles
+- Further improvements to user authentication forms.
+
+## Version 0.7.1
+
+- Updates requirements to Django 1.7.10
+
 ## Version 0.7
 
 - Adds `core.mixins`, that provide mixins for shared functionality, e.g. creating where clauses for user groups and subsets.
@@ -15,7 +27,10 @@
     - `geokey/users/templatetags/tags.py` > `geokey/users/templatetags/filter_tags.py`
 
 - Redesign of navigation elements
-- Fixes performance issues with counting number of files and comments per observation. Values are stored with the observation and updated each time either comment or file is added or removed. 
+- Fixes performance issues with counting number of files and comments per observation. Values are stored with the observation and updated each time either comment or file is added or removed.
+- Implements editing of lookup values
+- Bug fixes:
+    - [#279](https://github.com/ExCiteS/geokey/issues/279) Fixes validation of date fields; times before 10 a.m. used to cause an error.
 
 ## Version 0.6